fix(layout): Add InputGroup interface
Add an InputGroup interface, analogous to the InputNode. This InputGroup expects the leaves and
groups properties to be in the form of number[], instead of actual Node and Group objects. This
prevents the circular references between parent/child groups by allowing to specify leaves and
groups by index rather than full object.